Move sidebar surface contracts
