+++ Sources/QuillCodeApp/WorkspaceSurface.swift
@@ -167,426 +167,6 @@ public struct RuntimeDiagnosticSurface: Codable, Sendable, Hashable, Identifiabl
     }
 }
 
-public struct ProjectListSurface: Codable, Sendable, Hashable {
-    public var title: String
-    public var items: [ProjectItemSurface]
-    public var selectedProjectID: UUID?
-    public var emptyTitle: String
-
-    public init(
-        title: String = "Projects",
-        items: [ProjectItemSurface],
-        selectedProjectID: UUID?,
-        emptyTitle: String = "No projects yet"
-    ) {
-        self.title = title
-        self.items = items
-        self.selectedProjectID = selectedProjectID
-        self.emptyTitle = emptyTitle
-    }
-}
-
-public struct ProjectItemSurface: Codable, Sendable, Hashable, Identifiable {
-    public var id: UUID
-    public var name: String
-    public var path: String
-    public var connectionKindLabel: String
-    public var isRemote: Bool
-    public var actions: [ProjectItemActionSurface]
-    public var isSelected: Bool
-
-    public init(project: ProjectRef, selectedProjectID: UUID?) {
-        self.id = project.id
-        self.name = project.name
-        self.path = project.displayPath
-        self.connectionKindLabel = project.connection.kindLabel
-        self.isRemote = project.isRemote
-        self.actions = [
-            ProjectItemActionSurface(kind: .newChat, projectID: project.id),
-            ProjectItemActionSurface(kind: .refreshContext, projectID: project.id),
-            ProjectItemActionSurface(kind: .rename, projectID: project.id),
-            ProjectItemActionSurface(kind: .remove, projectID: project.id)
-        ]
-        self.isSelected = project.id == selectedProjectID
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case name
-        case path
-        case connectionKindLabel
-        case isRemote
-        case actions
-        case isSelected
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.id = try container.decode(UUID.self, forKey: .id)
-        self.name = try container.decode(String.self, forKey: .name)
-        self.path = try container.decode(String.self, forKey: .path)
-        self.connectionKindLabel = try container.decodeIfPresent(String.self, forKey: .connectionKindLabel) ?? "Local"
-        self.isRemote = try container.decodeIfPresent(Bool.self, forKey: .isRemote) ?? false
-        self.actions = try container.decodeIfPresent([ProjectItemActionSurface].self, forKey: .actions) ?? [
-            ProjectItemActionSurface(kind: .newChat, projectID: id),
-            ProjectItemActionSurface(kind: .refreshContext, projectID: id),
-            ProjectItemActionSurface(kind: .rename, projectID: id),
-            ProjectItemActionSurface(kind: .remove, projectID: id)
-        ]
-        self.isSelected = try container.decode(Bool.self, forKey: .isSelected)
-    }
-}
-
-public enum ProjectItemActionKind: String, Codable, Sendable, Hashable {
-    case newChat
-    case refreshContext
-    case rename
-    case remove
-
-    public var title: String {
-        switch self {
-        case .newChat:
-            return "New chat"
-        case .refreshContext:
-            return "Refresh context"
-        case .rename:
-            return "Rename"
-        case .remove:
-            return "Remove from list"
-        }
-    }
-}
-
-public struct ProjectItemActionSurface: Codable, Sendable, Hashable, Identifiable {
-    public var kind: ProjectItemActionKind
-    public var projectID: UUID
-    public var isEnabled: Bool
-    public var disabledReason: String?
-
-    public var id: String {
-        "\(projectID.uuidString)-\(kind.rawValue)"
-    }
-
-    public init(
-        kind: ProjectItemActionKind,
-        projectID: UUID,
-        isEnabled: Bool = true,
-        disabledReason: String? = nil
-    ) {
-        self.kind = kind
-        self.projectID = projectID
-        self.isEnabled = isEnabled
-        self.disabledReason = disabledReason
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case kind
-        case projectID
-        case isEnabled
-        case disabledReason
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.kind = try container.decode(ProjectItemActionKind.self, forKey: .kind)
-        self.projectID = try container.decode(UUID.self, forKey: .projectID)
-        self.isEnabled = try container.decodeIfPresent(Bool.self, forKey: .isEnabled) ?? true
-        self.disabledReason = try container.decodeIfPresent(String.self, forKey: .disabledReason)
-    }
-}
-
-public struct SidebarSurface: Codable, Sendable, Hashable {
-    public var title: String
-    public var items: [SidebarItemSurface]
-    public var selectedThreadID: UUID?
-    public var emptyTitle: String
-    public var isSelectionMode: Bool
-    public var selectedThreadIDs: Set<UUID>
-    public var selectionLabel: String
-    public var bulkActions: [SidebarBulkActionSurface]
-
-    public init(
-        title: String = "Chats",
-        items: [SidebarItemSurface],
-        selectedThreadID: UUID?,
-        emptyTitle: String = "No chats yet",
-        isSelectionMode: Bool = false,
-        selectedThreadIDs: Set<UUID> = [],
-        bulkActions: [SidebarBulkActionSurface] = []
-    ) {
-        self.title = title
-        self.items = items
-        self.selectedThreadID = selectedThreadID
-        self.emptyTitle = emptyTitle
-        self.isSelectionMode = isSelectionMode
-        self.selectedThreadIDs = selectedThreadIDs
-        self.selectionLabel = Self.selectionLabel(count: selectedThreadIDs.count)
-        self.bulkActions = bulkActions
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case title
-        case items
-        case selectedThreadID
-        case emptyTitle
-        case isSelectionMode
-        case selectedThreadIDs
-        case selectionLabel
-        case bulkActions
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.title = try container.decodeIfPresent(String.self, forKey: .title) ?? "Chats"
-        self.items = try container.decodeIfPresent([SidebarItemSurface].self, forKey: .items) ?? []
-        self.selectedThreadID = try container.decodeIfPresent(UUID.self, forKey: .selectedThreadID)
-        self.emptyTitle = try container.decodeIfPresent(String.self, forKey: .emptyTitle) ?? "No chats yet"
-        self.isSelectionMode = try container.decodeIfPresent(Bool.self, forKey: .isSelectionMode) ?? false
-        self.selectedThreadIDs = try container.decodeIfPresent(Set<UUID>.self, forKey: .selectedThreadIDs) ?? []
-        self.selectionLabel = try container.decodeIfPresent(String.self, forKey: .selectionLabel)
-            ?? Self.selectionLabel(count: self.selectedThreadIDs.count)
-        self.bulkActions = try container.decodeIfPresent([SidebarBulkActionSurface].self, forKey: .bulkActions) ?? []
-    }
-
-    public func filteredItems(matching query: String) -> [SidebarItemSurface] {
-        let normalizedQuery = query.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !normalizedQuery.isEmpty else {
-            return items
-        }
-        return items.filter { item in
-            let pinLabel = item.isPinned ? "pinned" : ""
-            let archivedLabel = item.isArchived ? "archived" : ""
-            return item.title.localizedCaseInsensitiveContains(normalizedQuery)
-                || item.subtitle.localizedCaseInsensitiveContains(normalizedQuery)
-                || item.searchText.localizedCaseInsensitiveContains(normalizedQuery)
-                || pinLabel.localizedCaseInsensitiveContains(normalizedQuery)
-                || archivedLabel.localizedCaseInsensitiveContains(normalizedQuery)
-        }
-    }
-
-    public var pinnedItems: [SidebarItemSurface] {
-        items.filter { $0.isPinned && !$0.isArchived }
-    }
-
-    public var recentItems: [SidebarItemSurface] {
-        items.filter { !$0.isPinned && !$0.isArchived }
-    }
-
-    public var archivedItems: [SidebarItemSurface] {
-        items.filter(\.isArchived)
-    }
-
-    private static func selectionLabel(count: Int) -> String {
-        switch count {
-        case 0:
-            return "No chats selected"
-        case 1:
-            return "1 chat selected"
-        default:
-            return "\(count) chats selected"
-        }
-    }
-}
-
-public struct SidebarItemSurface: Codable, Sendable, Hashable, Identifiable {
-    public var id: UUID
-    public var title: String
-    public var subtitle: String
-    public var searchText: String
-    public var actions: [SidebarItemActionSurface]
-    public var isSelected: Bool
-    public var isBulkSelected: Bool
-    public var isPinned: Bool
-    public var isArchived: Bool
-
-    public init(item: SidebarItem, selectedThreadID: UUID?, selectedThreadIDs: Set<UUID> = []) {
-        self.id = item.id
-        self.title = item.title
-        self.subtitle = item.subtitle
-        self.searchText = item.searchText
-        self.actions = Self.actions(for: item)
-        self.isSelected = item.id == selectedThreadID
-        self.isBulkSelected = selectedThreadIDs.contains(item.id)
-        self.isPinned = item.isPinned
-        self.isArchived = item.isArchived
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case title
-        case subtitle
-        case searchText
-        case actions
-        case isSelected
-        case isBulkSelected
-        case isPinned
-        case isArchived
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.id = try container.decode(UUID.self, forKey: .id)
-        self.title = try container.decode(String.self, forKey: .title)
-        self.subtitle = try container.decode(String.self, forKey: .subtitle)
-        self.searchText = try container.decode(String.self, forKey: .searchText)
-        self.actions = try container.decodeIfPresent([SidebarItemActionSurface].self, forKey: .actions) ?? []
-        self.isSelected = try container.decode(Bool.self, forKey: .isSelected)
-        self.isBulkSelected = try container.decodeIfPresent(Bool.self, forKey: .isBulkSelected) ?? false
-        self.isPinned = try container.decode(Bool.self, forKey: .isPinned)
-        self.isArchived = try container.decodeIfPresent(Bool.self, forKey: .isArchived) ?? false
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(id, forKey: .id)
-        try container.encode(title, forKey: .title)
-        try container.encode(subtitle, forKey: .subtitle)
-        try container.encode(searchText, forKey: .searchText)
-        try container.encode(actions, forKey: .actions)
-        try container.encode(isSelected, forKey: .isSelected)
-        try container.encode(isBulkSelected, forKey: .isBulkSelected)
-        try container.encode(isPinned, forKey: .isPinned)
-        try container.encode(isArchived, forKey: .isArchived)
-    }
-
-    private static func actions(for item: SidebarItem) -> [SidebarItemActionSurface] {
-        if item.isArchived {
-            return [
-                SidebarItemActionSurface(kind: .unarchive, threadID: item.id),
-                SidebarItemActionSurface(kind: .delete, threadID: item.id)
-            ]
-        }
-        return [
-            SidebarItemActionSurface(kind: .rename, threadID: item.id),
-            SidebarItemActionSurface(kind: .duplicate, threadID: item.id),
-            SidebarItemActionSurface(
-                kind: item.isPinned ? .unpin : .pin,
-                threadID: item.id
-            ),
-            SidebarItemActionSurface(kind: .archive, threadID: item.id),
-            SidebarItemActionSurface(kind: .delete, threadID: item.id)
-        ]
-    }
-}
-
-public enum SidebarBulkActionKind: String, Codable, Sendable, Hashable {
-    case select
-    case selectAll
-    case clearSelection
-    case pin
-    case unpin
-    case archive
-    case unarchive
-    case delete
-
-    public var title: String {
-        switch self {
-        case .select:
-            return "Select"
-        case .selectAll:
-            return "Select all"
-        case .clearSelection:
-            return "Done"
-        case .pin:
-            return "Pin"
-        case .unpin:
-            return "Unpin"
-        case .archive:
-            return "Archive"
-        case .unarchive:
-            return "Unarchive"
-        case .delete:
-            return "Delete"
-        }
-    }
-}
-
-public struct SidebarBulkActionSurface: Codable, Sendable, Hashable, Identifiable {
-    public var kind: SidebarBulkActionKind
-    public var commandID: String
-    public var title: String
-    public var isEnabled: Bool
-    public var isDestructive: Bool
-
-    public var id: String { commandID }
-
-    public init(
-        kind: SidebarBulkActionKind,
-        isEnabled: Bool = true,
-        isDestructive: Bool = false
-    ) {
-        self.kind = kind
-        self.commandID = Self.commandID(for: kind)
-        self.title = kind.title
-        self.isEnabled = isEnabled
-        self.isDestructive = isDestructive
-    }
-
-    public static func commandID(for kind: SidebarBulkActionKind) -> String {
-        switch kind {
-        case .select:
-            return "thread-selection-start"
-        case .selectAll:
-            return "thread-selection-select-all"
-        case .clearSelection:
-            return "thread-selection-clear"
-        case .pin:
-            return "thread-bulk-pin"
-        case .unpin:
-            return "thread-bulk-unpin"
-        case .archive:
-            return "thread-bulk-archive"
-        case .unarchive:
-            return "thread-bulk-unarchive"
-        case .delete:
-            return "thread-bulk-delete"
-        }
-    }
-}
-
-public enum SidebarItemActionKind: String, Codable, Sendable, Hashable {
-    case rename
-    case duplicate
-    case pin
-    case unpin
-    case archive
-    case unarchive
-    case delete
-
-    public var title: String {
-        switch self {
-        case .rename:
-            return "Rename"
-        case .duplicate:
-            return "Duplicate"
-        case .pin:
-            return "Pin"
-        case .unpin:
-            return "Unpin"
-        case .archive:
-            return "Archive"
-        case .unarchive:
-            return "Unarchive"
-        case .delete:
-            return "Delete"
-        }
-    }
-}
-
-public struct SidebarItemActionSurface: Codable, Sendable, Hashable, Identifiable {
-    public var kind: SidebarItemActionKind
-    public var threadID: UUID
-
-    public var id: String {
-        "\(threadID.uuidString)-\(kind.rawValue)"
-    }
-
-    public init(kind: SidebarItemActionKind, threadID: UUID) {
-        self.kind = kind
-        self.threadID = threadID
-    }
-}
-
 @MainActor
 public extension QuillCodeWorkspaceModel {
     func surface() -> WorkspaceSurface {
