+++ docs/DECISIONS.md
@@ -104,6 +104,7 @@
 - Thread archive is reversible and distinct from delete. Archived chats move to an Archived sidebar section, remain searchable, and can be unarchived back into the active thread list. Delete removes the thread from memory and the JSON thread store. Rename uses an explicit title field in the native UI, while duplicate creates an unpinned active `Copy:` thread that preserves the transcript and event audit.
 - Thread lifecycle mutations live in `WorkspaceThreadLifecycleEngine`. The workspace model owns stores, selected-project validation, terminal sync, and top-bar refresh, while the engine owns rename trimming, duplicate construction, pin toggles, single and bulk archive/unarchive state, delete removal, and newest-thread fallback selection.
 - Sidebar bulk selection is transient workspace state, not persisted thread metadata. `WorkspaceSidebarSelectionEngine` owns activation, clear, select-all, toggle, stale-ID pruning, and sidebar-order resolution. `WorkspaceSidebarBulkActionPlanner` owns selection-only command planning, visible-order target resolution, and post-mutation selection intent. `WorkspaceModel` stores selection state, applies the planned side effects, and persists thread/project changes. Rows expose `isBulkSelected` through the shared `WorkspaceSurface`, and bulk operations are command IDs (`thread-bulk-*`) just like single-thread actions.
+- Sidebar and project surface contracts live in `QuillCodeSidebarSurface.swift`. Project rows, thread rows, action labels, action IDs, bulk command IDs, search filtering, pinned/recent/archived grouping, and compatibility decoding stay beside the SwiftUI/HTML sidebar boundary; `WorkspaceSurface.swift` should only carry aggregate `projects` and `sidebar` payloads assembled from model state.
 - Sidebar command presentation has one source of truth in `QuillCodeSidebarCommandPresentation`. It owns visible primary/sidebar utility command ordering, display labels, native SF Symbols, HTML icon tokens, and Playwright test IDs so the SwiftUI sidebar and static HTML harness cannot drift as the Codex-like rail evolves.
 - Project row actions mirror thread row actions: the rail stays scannable, and New chat, Refresh context, Rename, and Remove from list live behind an ellipsis menu plus matching command-palette and slash routes. Removing a project forgets the project registry entry and detaches chats from that workspace root, but it never deletes files or thread transcripts.
 - Project connections are typed metadata, not path string conventions. Local projects use `Local`; registered SSH projects use `SSH Remote` everywhere in sidebar badges, command-palette titles, slash suggestions, and disabled-action explanations. SSH Remote rows can be registered with `/ssh user@host:/path` or `Project: Add SSH Remote...`. The first remote executor is noninteractive SSH: integrated-terminal commands, agent-authored `host.shell.run`, bounded `host.file.read`/`host.file.write`, `host.apply_patch`, `host.git.status`/`host.git.diff`, file/hunk `host.git.stage`/`host.git.restore`, `host.git.commit`, `host.git.push`, `host.git.pr.create`/`view`/`checks`/`checkout`/`comment`/`review`/`reviewers`/`labels`/`merge`, and `host.git.worktree.*` calls run through `ssh -T -o BatchMode=yes` from the selected remote root. SSH Remote projects intentionally offer the agent only remote-safe base tools (`host.shell.run`, bounded file read/write, apply-patch, git status/diff/stage/restore/commit/push/PR/worktree, plus app/plan/browser-style tools); environment tools are withheld until remote adapters exist, and unexpected local-only tool calls fail in the transcript instead of falling back to the desktop workspace.
